# doc/api.yaml
openapi: 3.0.3
...
info:
  title: WASAPhotoAPI
  version: 1.0.0
  description: |-
    API for the WASAPhoto application. <br>
    The API follows the CRUD philosophy, that maps HTTP methods TK
  contact:
    name: Emanuele Scaccia
    email: scaccia.2006749@studenti.uniroma1.it
    url: https://github.com/dev_excale/WASAPhoto
servers:
  - url: /
...
paths:
  /me/feed:
    get:
      operationId: getMyStream
      summary: Posts by the followed users
      description: Returns a list of posts made by the users the current user is following.
      tags: [ 'Post' ]
      security:
        - Bearer: [ ]
      responses:
        '200': # OK
          $ref: '#/components/responses/200-PostList'
        '400': # Bad Request
          $ref: '#/components/responses/400-BadRequest'
        '401': # Unauthorized
          $ref: '#/components/responses/401-Unauthorized'
        '500': # Internal Error
          $ref: '#/components/responses/500-InternalError'
    put:
      operationId: uploadPhoto
      summary: Add a post
      description: Uploads a new post to the user's feed.
      tags: [ 'Post' ]
      security:
        - Bearer: [ ]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: The webp image file to upload.
                  nullable: false
                json:
                  type: object
                  description: Information regarding the post
                  nullable: true
                  properties:
                    caption:
                      type: string
                      description: The caption of the post.
                      example: Working on this new project!
                      nullable: true
      responses:
        '201': # Created
          $ref: '#/components/responses/201-Post'
        '400': # Bad Request
          $ref: '#/components/responses/400-BadRequest'
        '401': # Unauthorized
          $ref: '#/components/responses/401-Unauthorized'
        '500': # Internal Error
          $ref: '#/components/responses/500-InternalError'
  /me/feed/{post_uuid}:
    delete:
      operationId: deletePhoto
      summary: Remove a post
      description: Removes a post from the user's feed.
      tags: [ 'Post' ]
      security:
        - Bearer: [ ]
      parameters:
        - $ref: '#/components/parameters/pathPostUUID'
      responses:
        '204': # No Content
          description: Post removed correctly
        '400': # Bad Request
          $ref: '#/components/responses/400-BadRequest'
        '401': # Unauthorized
          $ref: '#/components/responses/401-Unauthorized'
        '404': # Not Found
          $ref: '#/components/responses/404-PostNotFound'
        '500': # Internal Error
          $ref: '#/components/responses/500-InternalError'
...
  /users/{user_uuid}/feed:
    get:
      operationId: getUserFeed
      summary: Get all user's posts
      description: Returns the list of the posts made by the given user.
      tags: [ 'Post' ]
      security:
        - Bearer: [ ]
      parameters:
        - $ref: '#/components/parameters/pathUserUUID'
      responses:
        '200': # OK
          $ref: '#/components/responses/200-PostList'
        '401': # Unauthorized
          $ref: '#/components/responses/401-Unauthorized'
        '403': # Forbidden
          $ref: '#/components/responses/403-Banned'
        '404': # Not Found
          $ref: '#/components/responses/404-UserNotFound'
        '500': # Internal Error
          $ref: '#/components/responses/500-InternalError'
  /users/{user_uuid}/feed/{post_uuid}:
    get:
      operationId: getPhoto
      summary: Get a user's post
      description: Returns a given post made by the given user.
      tags: [ 'Post' ]
      security:
        - Bearer: [ ]
      parameters:
        - $ref: '#/components/parameters/pathUserUUID'
        - $ref: '#/components/parameters/pathPostUUID'
      responses:
        '200': # OK
          $ref: '#/components/responses/200-Post'
        '401': # Unauthorized
          $ref: '#/components/responses/401-Unauthorized'
        '403': # Forbidden
          $ref: '#/components/responses/403-Banned'
        '404': # Not Found
          $ref: '#/components/responses/404-PostNotFound'
        '500': # Internal Error
          $ref: '#/components/responses/500-InternalError'
...
components:
...
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer

  ##########################
  ## Parameter components ##
  ##########################
  parameters:
    pathUserUUID:
      name: user_uuid
      in: path
      required: true
      description: UUID of the target user
      example: 9a6a4415-9a50-4065-b345-8c22e8bb8c5f
      schema:
        $ref: '#/components/schemas/UUID'
    pathPostUUID:
      name: post_uuid
      in: path
      required: true
      description: UUID of the target post
      example: 9a6a4415-9a50-4065-b345-8c22e8bb8c5f
      schema:
        $ref: '#/components/schemas/UUID'
...
  schemas:
    UUID:
      type: string
      description: A UUID
      example: 9a6a4415-9a50-4065-b345-8c22e8bb8c5f
      format: uuid
...
    Post:
      type: object
      description: Information about a post
      properties:
        uuid:
          $ref: '#/components/schemas/UUID'
        username:
          type: string
          description: The username of the user who made the post.
          example: excale
        url_photo:
          type: string
          description: |-
            The URL of the photo. Note: all photos are encoded with webp format.
            Generally, the url will be of the form
            `https://api.example.com/v1/users/{username}/assets/{post_id}.webp`
          example: https://api.example.com/v1/users/excale/assets/3e270f31-6eaa-4418-ab97-7ed22f51d446.webp
        caption:
          type: string
          nullable: true
          description: The caption of the post.
          example: A day at the beach...
        num_likes:
          type: integer
          description: The number of likes the post has.
          example: 123
          minimum: 0
        num_comments:
          type: integer
          description: The number of comments the post has.
          example: 4
          minimum: 0
        timestamp:
          type: string
          description: The timestamp of the post, in `YYYY-MM-DD hh:mm:ss` format.
          example: 2021-01-01T00:00:00.000Z
    PostList:
      type: array
      description: A list of posts
      items:
        $ref: '#/components/schemas/Post'
...
  responses:
...
    200-Post:
      description: Post retrieved correctly.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Post'
    200-PostList:
      description: List retrieved correctly.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/PostList'
...
    201-Post:
      description: Post created correctly.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Post'
...
    400-BadRequest:
      description: The request was invalid or malformed (e.g. missing parameter).
    401-Unauthorized:
      description: The user is not logged in.
...
    403-Banned:
      description: The user has been banned by the given user.
    404-UserNotFound:
      description: The specified user doesn't exist.
    404-PostNotFound:
      description: The specified post doesn't exist.
...
    500-InternalError:
      description: An internal error occurred.